refactor: move cookie files to api/cookie/
Step 4.4 of source reorganization - move cookie-related files:
- JSCookie.cpp/h
- JSCookieMap.cpp/h
