+++ cmake/targets/BuildBun.cmake
@@ -225,8 +225,11 @@ set(BUN_ERROR_CODE_SCRIPT ${CWD}/src/codegen/generate-node-errors.ts)
 
 set(BUN_ERROR_CODE_SOURCES
   ${BUN_ERROR_CODE_SCRIPT}
+  ${CWD}/src/buntime/api/error
   ${CWD}/src/buntime/api/error/ErrorCode.ts
+  ${CWD}/src/buntime/api/error
   ${CWD}/src/buntime/api/error/ErrorCode.cpp
+  ${CWD}/src/buntime/api/error
   ${CWD}/src/buntime/api/error/ErrorCode.h
 )
 
@@ -591,11 +594,13 @@ foreach(i RANGE 0 ${BUN_OBJECT_LUT_SOURCES_MAX_INDEX})
 endforeach()
 
 WEBKIT_ADD_SOURCE_DEPENDENCIES(
+  ${CWD}/src/buntime/api/error
   ${CWD}/src/buntime/api/error/ErrorCode.cpp
   ${CODEGEN_PATH}/ErrorCode+List.h
 )
 
 WEBKIT_ADD_SOURCE_DEPENDENCIES(
+  ${CWD}/src/buntime/api/error
   ${CWD}/src/buntime/api/error/ErrorCode.h
   ${CODEGEN_PATH}/ErrorCode+Data.h
 )
@@ -868,6 +873,7 @@ target_include_directories(${bun} PRIVATE
   ${CWD}/src/buntime/api/console
   ${CWD}/src/buntime/api/inspector
   ${CWD}/src/buntime/api/error
+  ${CWD}/src/buntime/api/cookie
   ${CWD}/src/buntime/modules
   ${CWD}/src/js/builtins
   ${CWD}/src/napi
