+++ src/codegen/generate-node-errors.ts
@@ -1,5 +1,5 @@
 import path from "node:path";
-import NodeErrors from "../buntime/bindings/ErrorCode.ts";
+import NodeErrors from "../buntime/api/error/ErrorCode.ts";
 import { writeIfNotChanged } from "./helpers.ts";
 const outputDir = process.argv[2];
 
@@ -24,7 +24,7 @@ let zig = ``;
 enumHeader = `
 // clang-format off
 // Generated by: src/codegen/generate-node-errors.ts
-// Input:        src/buntime/bindings/ErrorCode.ts
+// Input:        src/buntime/api/error/ErrorCode.ts
 #pragma once
 
 #include <cstdint>
@@ -149,7 +149,7 @@ let builtindts = await Bun.file(builtindtsPath).text();
 
 let dts = `
 // Generated by: src/codegen/generate-node-errors.ts
-// Input:        src/buntime/bindings/ErrorCode.ts
+// Input:        src/buntime/api/error/ErrorCode.ts
 
 // Global error code functions for TypeScript
 `;
